Frontend: Clean up frontend test organization
